feat: modularize each actions
