+++ apps/frontend/acquisition-royale-home/src/features/mvp/categories.tsx
@@ -1,11 +1,11 @@
 import Acquire from './acquire'
-import Compete from './compete/Compete'
-import DepositRp from './deposit/DepositRp'
+import Compete from './compete'
+import DepositRp from './deposit'
 import Intern from './actions/Intern'
 import Rename from './actions/Rename'
 import Revive from './actions/Revive'
-import RPShop from './actions/RPShop'
-import WithdrawRp from './actions/WithdrawRp'
+import RPShop from './rpshop'
+import WithdrawRp from './withdraw/WithdrawRp'
 import Merge from './merge'
 import Trade from './trade'
 import { SubtabProps } from '../../components/Subtabs'
