+++ apps/frontend/acquisition-royale-home/src/features/mvp/compete/index.ts
@@ -1,0 +1,5 @@
+import Compete from './Compete'
+
+export { CompeteStore } from './CompeteStore'
+
+export default Compete
