+++ apps/frontend/acquisition-royale-home/src/features/mvp/deposit/index.ts
@@ -1,0 +1,4 @@
+import DepositRp from './DepositRp'
+
+export { DepositStore } from './DepositStore'
+export default DepositRp
