+++ apps/frontend/acquisition-royale-home/src/features/mvp/rpshop/RPShop.tsx
