+++ apps/frontend/acquisition-royale-home/src/features/mvp/rpshop/index.ts
@@ -1,0 +1,4 @@
+import RpShop from './RPShop'
+
+export { RpShopStore } from './RpShopStore'
+export default RpShop
